+++ src/mail/index.ts
@@ -1,0 +1,9 @@
+export {
+  MailtrapClient,
+  loadMailtrapConfig,
+  extractLinks,
+  firstLinkMatching,
+  type MailtrapConfig,
+  type MailtrapMessage,
+  type WaitForEmailOptions,
+} from './mailtrap.js';